fix: fix test failures and tax rate formatting
- Fix DocumentMailer tests to use EmailCollection correctly
- Fix EstimateToInvoiceConverter tests with proper dependencies
- Fix Model tests to use helper functions instead of direct creation
- Update tax_rate column to decimal(5,2) for fractional rates
- Fix tax rate display formatting in all views (18.00% -> 18%)
- All unit and feature tests now passing (101 passed, 2 skipped)
- Current test coverage: 30.8%

+++ database/migrations/2025_07_05_200317_create_invoice_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->text('description');
             $table->integer('quantity');
             $table->integer('unit_price');
-            $table->integer('tax_rate')->nullable();
+            $table->decimal('tax_rate', 5, 2)->nullable();
             $table->timestamps();
         });
     }
